requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.25
pymongo==3.12.1
